restructure subscriptions components
