+++ src/modules/subscriptions/index.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
-import Subscriptions from './subscriptions';
+import Subscriptions from './component';
 
 import { receiveQuery } from 'common/query/actions';
 import { receiveSetting } from 'common/setting/actions';
 import { fetchSubscriptions,
-         fetchSubscriptionUploads } from '../actions';
+         fetchSubscriptionUploads } from './actions';
 
 const mapStateToProps = ({ setting, subscriptions, user }) => ({
   subscriptions,
